--- module/Edm/src/Edm/TraitPartials/GetPublicVarsTrait.php
@@ -1,13 +1,0 @@
-<?php
-
-namespace Edm\TraitPartials;
-
-trait GetPublicVarsTrait {
-    public function getPublicVars () {
-        $class = $this; 
-        $vars = function() use ($class) { 
-            return get_object_vars($class); 
-        }; 
-        return $vars(); 
-    }
-}
