+++ module/Edm/src/Edm/Utils/GetPublicVarsTrait.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Edm\Utils;
+
+trait GetPublicVarsTrait {
+    public function getPublicVars () {
+        $class = $this; 
+        $vars = function() use ($class) { 
+            return get_object_vars($class); 
+        }; 
+        return $vars(); 
+    }
+}
