--- public/module-templates/edm-bbmjs/.gitignore
@@ -1,5 +1,0 @@
-.idea
-.sass-cache
-bower_components
-node_modules
-nbproject
