--- public/module-templates/edm-bbmjs/bower.json
@@ -1,30 +1,0 @@
-{
-  "name": "edm-bbmjs",
-  "version": "0.0.0",
-  "authors": [
-    "Ely De La Cruz <elycruz@elycruz.com>"
-  ],
-  "main": "distro/js/main.js",
-  "moduleType": [
-    "amd"
-  ],
-  "license": "MIT",
-  "ignore": [
-    "**/.*",
-    "node_modules",
-    "bower_components",
-    "test",
-    "tests"
-  ],
-  "dependencies": {
-    "common-classes": "elycruz/common-classes",
-    "requirejs": "~2.1.11",
-    "requirejs-text": "~2.0.12",
-    "sjljs": "elycruz/sjljs#~0.1.1",
-    "amplify": "~1.1.2",
-    "jquery": "~2.1.1",
-    "jquery-ui": "~1.10.4",
-    "modernizr": "~2.8.2",
-    "knockoutjs": "~3.1.0"
-  }
-}
