--- public/module-templates/edm-bbmjs/index.html
@@ -1,13 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head lang="en">
-    <meta charset="UTF-8">
-    <title></title>
-    <link rel="stylesheet" href="" />
-</head>
-<body>
-
-<script src="bower_components/requirejs/require.js"
-        data-main="src/edm-bbmjs/init"></script>
-</body>
-</html>
