--- public/module-templates/edm-bbmjs/package.json
@@ -1,14 +1,0 @@
-{
-  "name": "edm-bbmjs",
-  "version": "0.0.0",
-  "description": "",
-  "main": "distro/js/main.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "author": "",
-  "license": "ISC",
-  "devDependencies": {
-    "gulp": "^3.6.2"
-  }
-}
