--- public/module-templates/edm-bbmjs/sass/print.scss
@@ -1,3 +1,0 @@
-/* Welcome to Compass. Use this file to define print styles.
- * Import this file using the following HTML or equivalent:
- * <link href="/stylesheets/print.css" media="print" rel="stylesheet" type="text/css" /> */
