--- public/module-templates/edm-bbmjs/src/edm-bbmjs/lib/AbstractViewModel.js
@@ -1,8 +1,0 @@
-/**
- * Created by Ely on 5/24/2014.
- */
-define(['lib/AbstractModel'], function (AbstractModel) {
-    return AbstractModel.extend(function AbstractViewModel (options) {
-
-    });
-})
