--- public/module-templates/edm-bbmjs/src/edm-bbmjs/lib/utils/logger.js
@@ -1,10 +1,0 @@
-/**
- * Created by Ely on 5/24/2014.
- */
-define({
-    log: function () {
-        if (console) {
-            console.log.apply(console, sjl.argsToArray(arguments));
-        }
-    }
-});
