--- public/module-templates/edm-bbmjs/src/edm-bbmjs/main.js
@@ -1,8 +1,0 @@
-/**
- * Created by Ely on 5/24/2014.
- */
-define(['lib/utils/logger', 'lib/AbstractModel'], function (logger, AbstractModel) {
-    logger.log('main.js', '\n', 'line 5 of unminified source: "src/edm-bbmjs/main.js"');
-    var model = new AbstractModel();
-
-});
